Update 15
Changed the system of storing tasks in a file. Now every line does not
start with a number. Only tasks that are done start with [done] marker.
It should improve readability of the tasks file. Added error messages
for deleting selected task when nothing selected and for importing tasks
from file that does not exist. Added settings menu where user can change
the font size.